pyproject.toml
[project]
name = "spanish_moe_bert"
version = "0.1.0"
description = "Spanish MoE-BERT model with SPM tokenizer"
readme = "README.md"
requires-python = ">=3.9, <3.13"  # Changed: Avoid Python 3.13 for now
dependencies = [
    "torch==2.5.1",  # Changed: Fixed version instead of range
    "sentencepiece>=0.1.99",
    "sacremoses>=0.0.53",
    "datasets>=2.14.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "tqdm>=4.66.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "psutil>=5.9.0",
    "PyYAML>=6.0.1",
]

[project.scripts]
frankestein-transformer = "src.cli:main"

[project.optional-dependencies]
train = [
    "wandb>=0.15.0",
    "tensorboard>=2.14.0",
]
cuda = [
    "torch==2.5.1+cu118",  # CUDA 11.8 variant
]
cpu = [
    "torch==2.5.1+cpu",  # CPU-only variant
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# uv-specific configurations (optional)
